Update 19-manual-configuration.markdown
One omission found in all our documentation so far involves ignoring the origin of Liferay's properties.  We often just tell the reader that they can override a certain property in a portal-ext without explaining that it is possible to see all portal properties! 

This change would provide that information to the reader, but it is still considered buried important information.  I would suggest affording it its own page and explanation, including how to create their own custom portal-*.properties as well as Liferay instance-specific property files (as seen in the first 50 or so lines in the portal.properties file).

However, as that is a rather significant change I felt that would be rather presumptuous for me try and do! 

Please let me know what you think!