Update 14-installing-liferay-on-jboss-7.1.markdown
$JBOSS_HOME will only work on *nix-based systems.  For Windows it should be %JBOSS_HOME%, just like how JAVA_OPTS is called in the same line.

Also, there is no need to define JBOSS_HOME or use it to point to your server.policy.  Just providing the name of the server.policy file is sufficient if it is in the JBOSS_HOME/bin directory as the instructions indicated.

If these changes are not made JBoss will not start.